allow valid handshake returns for header auth

+++ src/Module/Api/ApiHandler.php
@@ -170,7 +170,11 @@ final class ApiHandler implements ApiHandlerInterface
         // send the version to API calls (this is used to determine return data for api4/api5)
         $input['api_version'] = $api_version;
 
-        // Create a simplified session for header authenticated sessions
+        /*
+         * Create a simplified session for header authenticated sessions
+         * If you are sending a handshake, then return a valid auth session.
+         * If you are doing anything else, you hide the session behind an MD5 hash of the username
+         */
         if (
             $header_auth &&
             $user instanceof User
@@ -178,19 +182,26 @@ final class ApiHandler implements ApiHandlerInterface
             $data             = [];
             $data['username'] = $user->username;
             $data['type']     = 'header';
-            $data['apikey']   = md5((string)$user->username);
             $data['value']    = $api_version;
-            // Session might not exist or has expired
-            if (!Session::read($data['apikey'])) {
-                Session::destroy($data['apikey']);
-                Session::create($data);
+            if ($is_handshake) {
+                // for a handshake there needs to be a valid auth response
+                $input['auth'] = Session::create($data);
+            } else {
+                $data['apikey'] = md5((string)$user->username);
+                // Session might not exist or has expired
+                if (!Session::read($data['apikey'])) {
+                    Session::destroy($data['apikey']);
+                    Session::create($data);
+                }
+
+                // Continue with the new session string to hide your header token
+                $input['auth'] = $data['apikey'];
             }
             if (in_array($api_version, Api::API_VERSIONS)) {
-                Session::write($data['apikey'], $api_version, $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::PERPETUAL_API_SESSION));
+                Session::write($input['auth'], $api_version, $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::PERPETUAL_API_SESSION));
             }
-            // Continue with the new session string to hide your header token
-            $input['auth'] = $data['apikey'];
         }
+
         // If it's not a handshake then we can allow it to take up lots of time
         if (!$is_handshake) {
             set_time_limit(0);
